Refactored slack
